Fix situation when _currentLevel is equal to _targetLevel and state machine switches to STATE_TARGETING.
Bad behaviour 1:
During a reset, if setLevel(0) is called, when reset finishes, state machine switches from IDLE to TARGETING and then into a RESETTING loop because  _currentLevel goes outside normal 0-100 range (to 255). Calling for any other level fixes the loop.
Evaluating _targetLevel == _currentLevel and asigning LEVEL_NONE to _targetLevel keeps the state machine in IDLE.
Bad behaviour 2:
If during an 100% end run calibration, setLevel(100) is called, after calibration finishes, the state machine goes into TARGETING and goes to 0% and then returning to 100%.
In setLevel function, calibration state is not evaluated and  _targetLevel is changed from LEVEL_NONE to 100. State machine switches from IDLE to TARGETING because of this. At this point direction is set to UP and machine moves to 0, IDLES but still having a target set, starts moving down to 100.
Calling a setLevel during this fixes the behaviour.